massively changes how i have been thinking about links/buttons rather than putting them in their own section, i'm bringing them into the single column at the very bottom.
this is introduced at the very beginning with the prominent login-to-update button.
then reinforced with the list of gists and actions at the bottom, and with every other page this architecture remains consistent

i've grouped the gistalt functionality (gistalt endpoints like /new, and the feedback button) on the left, while keeping github functionality on the right
This separation of concerns makes it easier to parse/use every page.